+++ projects/np-ui-lib/src/lib/np-auto-complete/np-auto-complete.component.ts
@@ -198,7 +198,7 @@ export class NpAutoCompleteComponent implements ControlValueAccessor, AfterViewI
   }
 
   _onKeydown(event: KeyboardEvent) {
-    if (event.key === 'Tab') {
+    if (event.key === 'Tab' || event.key === 'Escape') {
       this._close();
     }
   }