+++ projects/np-ui-lib/src/lib/np-dropdown/np-dropdown.component.ts
@@ -161,8 +161,12 @@ export class NpDropdownComponent implements ControlValueAccessor, AfterViewInit,
   }
 
   _onKeydown(event: KeyboardEvent) {
-    if (event.key === 'Tab') {
+    if (event.key === 'Tab' || event.key === 'Escape') {
       this._close();
+    }
+    if (event.key === 'ArrowDown') {
+      this._open();
+      event.preventDefault();
     }
   }