+++ projects/np-ui-lib/src/lib/np-tags/np-tags.component.ts
@@ -316,8 +316,12 @@ export class NpTagsComponent implements ControlValueAccessor, AfterViewInit, Aft
   }
 
   _onKeydown(event: KeyboardEvent) {
-    if (event.key === 'Tab') {
+    if (event.key === 'Tab' || event.key === 'Escape') {
       this._close();
+    }
+    if (event.key === 'ArrowDown') {
+      this._open();
+      event.preventDefault();
     }
   }