+++ projects/np-ui-lib/src/lib/np-time-picker/np-time-picker.component.ts
@@ -348,8 +348,12 @@ export class NpTimePickerComponent implements ControlValueAccessor, AfterViewIni
   }
 
   _onKeydown(event: KeyboardEvent) {
-    if (event.key === 'Tab') {
+    if (event.key === 'Tab' || event.key === 'Escape') {
       this._close();
+    }
+    if (event.key === 'ArrowDown') {
+      this._open();
+      event.preventDefault();
     }
   }