+++ src/app/np-number-box-demo/np-number-box-demo.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 
 import { NpNumberBoxDemoRoutingModule } from './np-number-box-demo-routing.module';
 import { NpNumberBoxDemoComponent } from './np-number-box-demo.component';
-import { NpNumberBoxModule, NpTabsModule, NpCardModule } from 'np-ui-lib';
+import { NpNumberBoxModule, NpTabsModule, NpCardModule, NpAlertModule } from 'np-ui-lib';
 import { FormsModule } from '@angular/forms';
 
 @NgModule({
@@ -14,7 +14,8 @@ import { FormsModule } from '@angular/forms';
     NpNumberBoxModule,
     FormsModule,
     NpTabsModule,
-    NpCardModule
+    NpCardModule,
+    NpAlertModule
   ]
 })
 export class NpNumberBoxDemoModule { }